# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "2Gather API DOC",
		"description": "OpenAPI Specification for 2Gather API"
	},
	"schemes": [
		"http"
	],
	"host": "localhost:4003",
	"basePath": "/api",
...
	"paths": {
...
		"/user": {
			"get": {
				"tags": [
					"User"
				],
				"summary": "Get user profile",
				"description": "Get user profile functionality for users.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/languageInHeader"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"put": {
				"tags": [
					"User"
				],
				"summary": "Update user profile",
				"description": "Update user profile functionality for all",
				"consumes": "application/x-www-form-urlencoded",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/languageInHeader"
					},
					{
						"name": "firstName",
						"description": "User first name, required(For user role)",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "lastName",
						"description": "User last name, required(For user role)",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "birthDate",
						"description": "User birth date(For user role)",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "gender",
						"description": "User gender(For user user role)",
						"type": "number",
						"in": "formData",
						"required": false
					},
					{
						"name": "phoneNumber",
						"description": "User's phone number(For both roles)",
						"type": "number",
						"in": "formData",
						"required": false
					},
					{
						"name": "countryId",
						"description": "User's county id(For both roles)",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "fullName",
						"description": "Corporate full name, required(For corporate role)",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "taxNumber",
						"description": "Corporate tax number, required(For corporate role)",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "contactPerson",
						"description": "Corporate contact person(For corporate role)",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "deleteAvatar",
						"description": "Sending true means user deleted avatar",
						"type": "boolean",
						"in": "formData",
						"required": false
					},
					{
						"name": "avatar",
						"description": "User avatar(For both roles)",
						"type": "file",
						"in": "formData",
						"required": false
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/avatar": {
			"put": {
				"tags": [
					"User"
				],
				"summary": "Upload user avatar(FOR WEB ONLY)",
				"description": "Upload user avatar(FOR WEB ONLY) functionality for all",
				"consumes": "application/x-www-form-urlencoded",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"name": "avatar",
						"description": "User avatar(For both roles)",
						"type": "file",
						"in": "formData",
						"required": false
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/password": {
			"put": {
				"tags": [
					"User"
				],
				"summary": "Set user password",
				"description": "Functionality for users to set new or change old password.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"in": "body",
						"name": "Set password",
						"description": "Set password request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"oldPassword": {
									"type": "string"
								},
								"newPassword": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/adminList": {
			"post": {
				"tags": [
					"User"
				],
				"summary": "Get user list for admin",
				"description": "Functionality to get user list for admin.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Get user list",
						"description": "Get user list body.",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"pageNo": {
									"type": "number"
								},
								"limit": {
									"type": "number"
								},
								"search": {
									"type": "string"
								},
								"country": {
									"type": "string"
								},
								"role": {
									"type": "number"
								},
								"status": {
									"type": "number"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/block": {
			"put": {
				"tags": [
					"User"
				],
				"summary": "Block or activate user",
				"description": "Block or activate user functionality for admins.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/idInQuery"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/count": {
			"post": {
				"tags": [
					"User"
				],
				"summary": "Get user count by filters for admin",
				"description": "Functionality to get user count by filters for admin.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Get user count",
						"description": "Get user count body.",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"filters": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"search": {
												"type": "string"
											},
											"country": {
												"type": "string"
											},
											"role": {
												"type": "number"
											},
											"status": {
												"type": "number"
											}
										}
									}
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/adminDetails": {
			"get": {
				"tags": [
					"User"
				],
				"summary": "Get user details",
				"description": "Get user details functionality for admins.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/idInQuery"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
	},
	"parameters": {
		"Authorization": {
			"name": "Authorization",
			"in": "header",
			"required": true,
			"description": "Bearer token",
			"type": "string"
		},
...
		"idInQuery": {
			"name": "id",
			"in": "query",
			"required": true,
			"description": "Item Id",
			"type": "string"
		},
...
		"languageInHeader": {
			"name": "language",
			"in": "header",
			"required": true,
			"description": "Language",
			"type": "number"
		}
	}
}